style: make landing at once
