+++ src/pages/landing/index.tsx
@@ -1,51 +1,11 @@
 import Layout from '../../layouts/Layout';
-import { useEffect, useState } from 'react';
-import FirstLanding from './_first';
-import SecondLanding from './_second';
+import useWindowSize from '../../hooks/useWindowSize';
+import MobileLanding from './_mobile';
+import PCLanding from './_pc';
 
 const Landing = () => {
-    const [scrollDirection, setScrollDirection] = useState(false);
-    const [isFirst, setIsFirst] = useState(true);
-    const [count, setCount] = useState(0);
-
-    useEffect(() => {
-        const threshold = 0;
-        let lastScrollY = window.pageYOffset;
-        let ticking = false;
-
-        const updateScrollDirection = () => {
-            const scrollY = window.pageYOffset;
-
-            if (Math.abs(scrollY - lastScrollY) < threshold) {
-                ticking = false;
-                return;
-            }
-            setScrollDirection(scrollY > lastScrollY ? true : false);
-            if (count == 0) {
-                if (scrollDirection) {
-                    setIsFirst(false);
-                    setCount(prev => prev + 1);
-                }
-            }
-            //setIsFirst(scrollDirection ? true : false);
-            lastScrollY = scrollY > 0 ? scrollY : 0;
-            ticking = false;
-        };
-
-        const onScroll = () => {
-            if (!ticking) {
-                window.requestAnimationFrame(updateScrollDirection);
-                ticking = true;
-            }
-        };
-
-        window.addEventListener('scroll', onScroll);
-
-        return () => window.removeEventListener('scroll', onScroll);
-    }, [scrollDirection]);
-
-    // return <div>{scrollDirection ? <SecondLanding /> : <FirstLanding />}</div>
-    return <div>{scrollDirection ? <SecondLanding /> : <FirstLanding isFirst={isFirst} />}</div>;
+    const size = useWindowSize();
+    return <div>{size.width && size.width > 640 ? <PCLanding /> : <MobileLanding />}</div>;
 };
 
 Landing.Layout = Layout;
